requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.59.0
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-model-utils==4.3.1
django-notifications-hq==1.8.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
jsonfield==3.1.0
mysqlclient==2.2.1
oauthlib==3.2.2
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
swapper==1.3.0
tzdata==2023.3
urllib3==2.1.0
